refactor: 불필요한 import 삭제 및 패키지명 변경
