+++ src/main/java/chess/dto/StatusDto.java
@@ -1,6 +1,8 @@
 package chess.dto;
 
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class StatusDto {
